@@ -1,15 +1,26 @@
-import { Component, OnInit } from '@angular/core';
-import { TranslateService } from '@ngx-translate/core';
-import { environment } from 'src/environments/environment';
+import { Router } from "@angular/router";
+import { Component, OnInit } from "@angular/core";
+import { TranslateService } from "@ngx-translate/core";
+import { environment } from "src/environments/environment";
+import { AuthenticationService } from "./home/auth.service";
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
+  selector: "app-root",
+  templateUrl: "./app.component.html",
+  styleUrls: ["./app.component.css"],
 })
 export class AppComponent implements OnInit {
-  constructor(private translateService: TranslateService) {}
+  constructor(
+    private translateService: TranslateService,
+    private authenticationService: AuthenticationService,
+    private router: Router
+  ) {}
   ngOnInit(): void {
     this.translateService.use(environment.defaultLanguage);
+    this.authenticationService.getUser().subscribe(() => {
+      if (this.authenticationService.isAuthenticated) {
+        this.router.navigate(["/employee"]);
+      }
+    });
   }
 }
