@@ -3,10 +3,15 @@ export interface IUser {
   name: string;
   date: string;
   imageUrl: string;
-  location?: {
-    address: string;
+  address?: {
+    address1: string;
+    address2: string;
     city: string;
     country: string;
+    fax: string;
+    phone: string;
+    postalCode: string;
+    stateOrProvince: string;
   };
   isAdmin: boolean;
 }
