@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using DocuSign.eSign.Api;
 using DocuSign.eSign.Client;
 using Microsoft.AspNetCore.Authentication;
@@ -8,16 +9,30 @@ namespace DocuSign.MyHR.Services
 {
     public class DocuSignApiProvider : IDocuSignApiProvider
     {
-        private readonly Configuration _docuSignConfig;
-        private Lazy<IUsersApi> _usersApi => new Lazy<IUsersApi>(() => new UsersApi(_docuSignConfig));
+        private readonly IHttpContextAccessor _httpContextAccessor;
+
+        private Lazy<IUsersApi> _usersApi => new Lazy<IUsersApi>(() => new UsersApi(_docuSignConfig.Value));
+        private Lazy<IEnvelopesApi> _envelopApi => new Lazy<IEnvelopesApi>(() => new EnvelopesApi(_docuSignConfig.Value));
+
+        private Lazy<Configuration> _docuSignConfig => new Lazy<Configuration>(() =>
+            {
+                var apiClient = new ApiClient(Context.Account.BaseUri + "/restapi");
+                var docuSignConfig = new Configuration(apiClient);
+                var token = _httpContextAccessor.HttpContext.GetTokenAsync("access_token").Result;
+                if (string.IsNullOrEmpty(token))
+                {
+                    token = _httpContextAccessor.HttpContext.User.FindAll("access_token").First().Value;
+                }
+                docuSignConfig.AddDefaultHeader("Authorization", "Bearer " + token);
+                return docuSignConfig;
+            });
 
         public DocuSignApiProvider(IHttpContextAccessor httpContextAccessor)
         {
-            var apiClient = new ApiClient(Context.Account.BaseUri + "/restapi");
-            _docuSignConfig = new Configuration(apiClient);
-            _docuSignConfig.AddDefaultHeader("Authorization", "Bearer " + httpContextAccessor.HttpContext.GetTokenAsync("access_token").Result);
-        } 
+            _httpContextAccessor = httpContextAccessor;
+        }
 
         public IUsersApi UsersApi => _usersApi.Value;
+        public IEnvelopesApi EnvelopApi => _envelopApi.Value;
     }
 }
